Move alpha modulation into paint conversion (Step 1)
Skia has a rule that says (roughly), any SkShader is modulated by the
paint alpha. The CPU backend implements this precisely - the entire
shader sees the opaque paint color, and the alpha is applied once at
the root. The GPU backend has previously made each SkShader responsible
for doing this modulation. That led to two kinds of problems:

  1) Some shaders forgot to do the modulation at all.
  2) In shader trees, it's possible for the alpha to get applied
     multiple times, or to child effects where it isn't desired.

A consequence of #2 is that the GPU implementation of things like
blend-shaders have to jump through hoops to invoke their children with
opaque versions of their input colors.

This CL allows us to remove the explicit alpha modulation from a variety
of shaders, and should also cleanup of the blend shader logic. However,
to get the CL past chrome's layout tests, we need to guard the change.
So this version just makes the major change so paint conversion, in a
way that will work universally. Once chrome switches to the new method,
it will be safe to clean up the various shader implementations.

Bug: skia:11942
Change-Id: I518534becdd5d10cbd78cf3ff7d4a46dd1faabf9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/464116
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>