Keep uniform types even in release.
We need to know a uniform's type to convert it to the uniform-buffer's
expected format when uniforms are set. We have this data in other
arrays, but it's not accessible to the program data manager.

We don't necessarily need to use a bitfield here, if we think there's
ever a possibility of having a uniform buffer > 16MB in size.

Change-Id: I88b3936de5319d65c76891ec0f1d87badc71d2dd
Bug: skia:12339
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/465076
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>