Better handling of coverage-as-alpha during paint conversion
This is generally not a property of individual FPs in the shader tree,
it's a property of the root level FP, and how it applies alpha to the
color generated by the shader FP. So, two improvements:

  1) Remove all of the coverage-as-alpha inhibition happening in
     individual shaders.
  2) At the root of paint conversion, if there's a shader FP (and no
     primitive color), we either: use ApplyPaintAlpha, which *is* safe
     with coverage-as-alpha (updated flags to reflect this), or disable
     coverage as alpha (because our FP is now just generating a color and
     ignoring incoming alpha).

Note that we *could* choose to ALWAYS use ApplyPaintAlpha, even if the
paint color is opaque. That could result in extra (useless) work in some
situations, but it would enable the coverage-as-alpha optimization far
more often than we get it today.

Also: We don't change the behavior when there is a primitive color,
because that code path is already disabling coverage-as-alpha, thanks to
OverrideInput.

Change-Id: I1f040ee1217cf03e731ad1b8b497dc8e4df2de2b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/469518
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>